+++ R/serialize.R
@@ -25,9 +25,29 @@ unmarshall <- function(raw_vec) {
 }
 
 
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Serialize an R object to a raw vector, while minimising memory allocations
+#'
+#' This is a slightly more complex version of \code{marshall()} which
+#' pre-calculates the size of the serialized representation.  This means there
+#' only needs to be 1 memory allocation (not including those internal to R),
+#' and gives a speed boost for medium-to-large objects
+#'
+#' @param robj R object
+#'
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+marshall_minimize_malloc <- function(robj) {
+  .Call(marshall_minimize_malloc_, robj)
+}
+
+
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #' Calculate the size of the R object when it is serialized
+#'
+#' This function goes through all the motions of serializing an object, but
+#' does nothing with the bytes other than to tally the total length.
 #'
 #' @param robj R object
 #'