+++ src/calc-size-only.h
@@ -1,0 +1,2 @@
+
+int calc_marshalled_size(SEXP robj);