+++ NEWS.md
@@ -1,3 +1,9 @@
+# serializer 0.1.3   2020-09-22
+
+* Features: `marshall_minimize_malloc()` pre-calculates the resulting size of 
+  the serialized representation. This minimizes memory allocations and 
+  leads to faster serialization for medium to large objects
+
 # serializer 0.1.2   2020-09-22
 
 * Feature: `calc_marshalled_size()` added. This function will only calculate